requirements.txt
google-cloud-vision
Flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
google-generativeai==0.8.5
gunicorn==21.2.0

